Fixed merge conflicts within web.config file

+++ CSULB.GetUsGrub/CSULB.GetUsGrub.Models/Models/Interfaces/IMenuItem.cs
@@ -4,7 +4,7 @@
     {
         string ItemName { get; set; }
 
-        double ItemPrice { get; set; }
+        decimal ItemPrice { get; set; }
 
         string ItemPicture { get; set; }
